# lainnya/asisten_mahasiswa_openapi.yaml
openapi: 3.0.3
info:
  title: Asisten Mahasiswa API
  version: 1.0.0
  description: >
    Kontrak API untuk jadwal akademik, ujian, event, dan layanan helpdesk (SOP ULT).
servers:
  - url: https://api.your-campus.local/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local Dev
...
paths:
...
  /services:
    get:
      tags: [Helpdesk]
      summary: Daftar layanan SOP ULT
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
    post:
      tags: [Helpdesk]
      summary: Buat layanan (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
  /services/{id}:
    get:
      tags: [Helpdesk]
      summary: Detail layanan (syarat, alur, produk)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceDetail" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Service:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        unit_owner: { type: string }
        sla_days: { type: integer }
        fee_rp: { type: integer }
        is_active: { type: boolean }
        sop_ref: { type: string }
    ServiceCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        unit_owner: { type: string }
        sla_days: { type: integer }
        fee_rp: { type: integer }
        is_active: { type: boolean }
        sop_ref: { type: string }
    Requirement:
      type: object
      properties:
        item: { type: string }
        is_mandatory: { type: boolean }
        note: { type: string, nullable: true }
        sort_no: { type: integer }
    Workflow:
      type: object
      properties:
        step_no: { type: integer }
        description: { type: string }
        link_url: { type: string, nullable: true }
    ServiceDetail:
      type: object
      properties:
        service: { $ref: "#/components/schemas/Service" }
        requirements:
          type: array
          items: { $ref: "#/components/schemas/Requirement" }
        workflows:
          type: array
          items: { $ref: "#/components/schemas/Workflow" }
        documents:
          type: array
          items:
            type: object
            properties:
              product_name: { type: string }
...
security:
  - bearerAuth: []